bugfix: fixed warning...

+++ src/chess/board.cpp
@@ -97,7 +97,7 @@ Board Board::load_from_fen(std::string fen)
     board.next = sections[1] == "w" ? Team::White : Team::Black;
 
     auto castling_rights = 0;
-    for (int i = 0; i < sections[2].length(); ++i) {
+    for (int i = 0; i < (int)sections[2].length(); ++i) {
         auto current = sections[2].substr(i, i + 1).c_str()[0];
         switch (current) {
         case 'K':
